Changed game ID from long to an int.

+++ src/main/java/com/jakethomson/tictactoe/models/Game.java
@@ -1,12 +1,9 @@
 package com.jakethomson.tictactoe.models;
 
-import org.springframework.stereotype.Service;
-
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import java.util.concurrent.atomic.AtomicLong;
 
 /** Game is a POJO that holds all data on a single game. */
 @Entity
@@ -15,7 +12,7 @@ public class Game {
     // A unique ID is generated each time a new game is instantiated.
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private long id;
+    private int id;
     private String player_x_id;
     private String player_o_id;
     private String[] board;
@@ -38,7 +35,7 @@ public class Game {
     }
 
     // Getters and status/board setter.
-    public long getId() { return id; }
+    public int getId() { return id; }
     public String getPlayer_x_id() { return player_x_id; }
     public String getPlayer_o_id() { return player_o_id; }
     public String[] getBoard() { return board; }
